requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
backports.zoneinfo==0.2.1
boto3==1.26.64
botocore==1.29.64
decorator==5.1.1
dj-database-url==1.2.0
Django==4.1.6
django-cors-headers==3.13.0
django-storages==1.13.2
djangorestframework==3.14.0
executing==1.2.0
gunicorn==20.1.0
ipython==8.9.0
jedi==0.18.2
jmespath==1.0.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.36
psycopg2==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.9.0
urllib3==1.26.14
wcwidth==0.2.6
